upgrade file-stream-rotator package
The documentation for this package indicates support for the
capabilities of the file-stream-rotator package for version 1.0.0.
When I attempted to use the frequence based rotation it wasn't working.
Looking at the docs for the file-stream-rotator package version
currently used:
https://www.npmjs.com/package/file-stream-rotator/v/0.6.1
The version being used does not suppor the frequency options documented,
however, the 1.0.0 version does.
https://www.npmjs.com/package/file-stream-rotator
This will align the functionality with what is currently documented, I
tested locally with an up to date version of the file-stream-rotator
package and it worked.